@@ -4,41 +4,42 @@ publish_to: "none"
 version: 0.0.1+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  stacked: ^2.0.2
-  auto_route: ^1.0.2
-  get_it: ^6.0.0
-  injectable: ^1.2.2
+  stacked: ^3.4.2
+  auto_route: ^7.8.4
+  get_it: ^7.6.7
+  injectable: ^2.3.2
 
-  firebase_core: ^1.0.3
-  firebase_auth: ^1.0.2
-  cloud_firestore: ^1.0.4
-  firebase_storage: ^8.0.2
-  firebase_messaging: ^9.1.1
+  firebase_core: ^2.25.5
+  firebase_auth: ^4.17.6
+  cloud_firestore: ^4.15.6
+  firebase_storage: ^11.6.7
+  firebase_messaging: ^14.7.17
 
-  flash: ^1.5.1
+  flash: ^3.1.0
 
-  syncfusion_flutter_pdfviewer: ^19.1.55-beta
+  syncfusion_flutter_pdfviewer: ^24.2.8
 
-  url_launcher: ^6.0.3
-  font_awesome_flutter: ^9.0.0
-  intl: ^0.17.0
-  image_picker: ^0.7.4
+  url_launcher: ^6.2.5
+  font_awesome_flutter: ^10.7.0
+  intl: ^0.19.0
+  image_picker: ^1.0.7
 
+  meta: any
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^1.12.2
-  auto_route_generator: ^1.0.2
-  injectable_generator: ^1.2.2
+  build_runner: ^2.4.8
+  auto_route_generator: ^7.3.2
+  injectable_generator: ^2.4.1
 
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.13.1
 
 flutter_icons:
   android: "launcher_icon"
