@@ -1,11 +1,9 @@
-import "package:meta/meta.dart";
-
 class KError {
   final String errorCode;
   final String userFriendlyMessage;
 
   KError({
-    @required this.errorCode,
-    @required this.userFriendlyMessage,
+    required this.errorCode,
+    required this.userFriendlyMessage,
   });
 }
