@@ -3,8 +3,8 @@ import 'package:injectable/injectable.dart';
 
 @lazySingleton
 class FirebaseService {
-  FirebaseApp _firebaseApp;
-  FirebaseApp get firebaseApp => _firebaseApp;
+  FirebaseApp? _firebaseApp;
+  FirebaseApp? get firebaseApp => _firebaseApp;
 
   Future<void> init() async {
     _firebaseApp = await Firebase.initializeApp();
