@@ -11,9 +11,9 @@ class FirestorageService {
   final AuthService _authService = locator<AuthService>();
   final FirebaseStorage _firebaseStorage = FirebaseStorage.instance;
 
-  TaskSnapshot _taskSnapshot;
-  TaskSnapshot get taskSnapshot => _taskSnapshot;
-  void _setTaskSnapshot(TaskSnapshot newValue) {
+  TaskSnapshot? _taskSnapshot;
+  TaskSnapshot? get taskSnapshot => _taskSnapshot;
+  void _setTaskSnapshot(TaskSnapshot? newValue) {
     _taskSnapshot = newValue;
   }
 
@@ -21,7 +21,7 @@ class FirestorageService {
     try {
       _setTaskSnapshot(null);
       final Reference reference = _firebaseStorage.ref(
-        "profile_images/${_authService.currentUser.id}/${DateTime.now().toIso8601String()}.jpeg",
+        "profile_images/${_authService.currentUser!.id}/${DateTime.now().toIso8601String()}.jpeg",
       );
       final TaskSnapshot _taskSnapshotTemp = await reference.putFile(
         profileImage,
