@@ -2,14 +2,9 @@ import "package:flash/flash.dart";
 import "package:flutter/material.dart";
 
 class FlashHelper {
-  static Color _backgroundColor(BuildContext context) {
+  static TextStyle _titleStyle(BuildContext context, [Color? color]) {
     final ThemeData theme = Theme.of(context);
-    return theme.dialogTheme.backgroundColor ?? theme.dialogBackgroundColor;
-  }
-
-  static TextStyle _titleStyle(BuildContext context, [Color color]) {
-    final ThemeData theme = Theme.of(context);
-    return (theme.dialogTheme.titleTextStyle ?? theme.textTheme.headline6)
+    return (theme.dialogTheme.titleTextStyle ?? theme.textTheme.titleLarge)
             ?.copyWith(color: color) ??
         const TextStyle(
           fontSize: 21.0,
@@ -17,9 +12,9 @@ class FlashHelper {
         );
   }
 
-  static TextStyle _contentStyle(BuildContext context, [Color color]) {
+  static TextStyle _contentStyle(BuildContext context, [Color? color]) {
     final ThemeData theme = Theme.of(context);
-    return (theme.dialogTheme.contentTextStyle ?? theme.textTheme.bodyText2)
+    return (theme.dialogTheme.contentTextStyle ?? theme.textTheme.bodyMedium)
             ?.copyWith(color: color) ??
         const TextStyle(
           fontSize: 15.0,
@@ -27,10 +22,10 @@ class FlashHelper {
         );
   }
 
-  static Future<T> successBar<T>(
+  static Future<T?> successBar<T>(
     BuildContext context, {
-    String title,
-    @required String message,
+    String? title,
+    required String message,
     Duration duration = const Duration(seconds: 3),
   }) {
     return showFlash<T>(
@@ -42,9 +37,12 @@ class FlashHelper {
       ) {
         return Flash<T>(
           controller: controller,
-          horizontalDismissDirection: HorizontalDismissDirection.horizontal,
-          backgroundColor: Colors.black87,
+          dismissDirections: [
+            FlashDismissDirection.endToStart,
+            FlashDismissDirection.startToEnd,
+          ],
           child: FlashBar(
+            controller: controller,
             title: title == null
                 ? null
                 : Text(
@@ -54,7 +52,7 @@ class FlashHelper {
                       Colors.white,
                     ),
                   ),
-            message: Text(
+            content: Text(
               message,
               style: _contentStyle(
                 context,
@@ -65,17 +63,17 @@ class FlashHelper {
               Icons.check_circle,
               color: Colors.blue[300],
             ),
-            leftBarIndicatorColor: Colors.blue[300],
+            indicatorColor: Colors.blue[300],
           ),
         );
       },
     );
   }
 
-  static Future<T> errorBar<T>(
+  static Future<T?> errorBar<T>(
     BuildContext context, {
-    String title,
-    @required String message,
+    String? title,
+    required String message,
     Duration duration = const Duration(seconds: 3),
   }) {
     return showFlash<T>(
@@ -87,9 +85,12 @@ class FlashHelper {
       ) {
         return Flash<T>(
           controller: controller,
-          horizontalDismissDirection: HorizontalDismissDirection.horizontal,
-          backgroundColor: Colors.black87,
+          dismissDirections: [
+            FlashDismissDirection.endToStart,
+            FlashDismissDirection.startToEnd,
+          ],
           child: FlashBar(
+            controller: controller,
             title: title == null
                 ? null
                 : Text(
@@ -99,7 +100,7 @@ class FlashHelper {
                       Colors.white,
                     ),
                   ),
-            message: Text(
+            content: Text(
               message,
               style: _contentStyle(
                 context,
@@ -110,17 +111,17 @@ class FlashHelper {
               Icons.warning,
               color: Colors.red[300],
             ),
-            leftBarIndicatorColor: Colors.red[300],
+            indicatorColor: Colors.red[300],
           ),
         );
       },
     );
   }
 
-  static Future<T> infoBar<T>(
+  static Future<T?> infoBar<T>(
     BuildContext context, {
-    String title,
-    @required String message,
+    String? title,
+    required String message,
     Duration duration = const Duration(seconds: 3),
   }) {
     return showFlash<T>(
@@ -132,9 +133,12 @@ class FlashHelper {
       ) {
         return Flash<T>(
           controller: controller,
-          horizontalDismissDirection: HorizontalDismissDirection.horizontal,
-          backgroundColor: Colors.black87,
+          dismissDirections: [
+            FlashDismissDirection.endToStart,
+            FlashDismissDirection.startToEnd,
+          ],
           child: FlashBar(
+            controller: controller,
             title: title == null
                 ? null
                 : Text(
@@ -144,7 +148,7 @@ class FlashHelper {
                       Colors.white,
                     ),
                   ),
-            message: Text(
+            content: Text(
               message,
               style: _contentStyle(
                 context,
@@ -155,45 +159,7 @@ class FlashHelper {
               Icons.info_outline,
               color: Colors.green[300],
             ),
-            leftBarIndicatorColor: Colors.green[300],
-          ),
-        );
-      },
-    );
-  }
-
-  static Future<T> simpleDialog<T>(
-    BuildContext context, {
-    @required String title,
-    @required String message,
-    Color messageColor,
-    List<Widget> actions,
-  }) {
-    return showFlash<T>(
-      context: context,
-      persistent: false,
-      builder: (
-        BuildContext context,
-        FlashController<T> controller,
-      ) {
-        return Flash<T>.dialog(
-          controller: controller,
-          backgroundColor: _backgroundColor(context),
-          margin: const EdgeInsets.only(
-            left: 40.0,
-            right: 40.0,
-          ),
-          borderRadius: const BorderRadius.all(Radius.circular(8.0)),
-          child: FlashBar(
-            title: Text(
-              title,
-              style: _titleStyle(context),
-            ),
-            message: Text(
-              message,
-              style: _contentStyle(context, messageColor),
-            ),
-            actions: actions,
+            indicatorColor: Colors.green[300],
           ),
         );
       },
