@@ -1,18 +1,19 @@
+import 'package:arrowad_grade_eleven/src/app/models/k_homework_item.dart';
+import 'package:arrowad_grade_eleven/src/app/utils/colors.dart';
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart' as intl;
 import 'package:stacked/stacked.dart';
 
-import 'package:arrowad_grade_eleven/src/app/models/k_homework_item.dart';
-import 'package:arrowad_grade_eleven/src/app/utils/colors.dart';
-
 import './homework_item_info_view_model.dart';
 
+@RoutePage()
 class HomeworkItemInfoView extends StatelessWidget {
   final KHomeworkItem homeworkItem;
 
   const HomeworkItemInfoView({
-    Key key,
-    @required this.homeworkItem,
+    Key? key,
+    required this.homeworkItem,
   }) : super(key: key);
 
   @override
@@ -22,7 +23,7 @@ class HomeworkItemInfoView extends StatelessWidget {
       builder: (
         BuildContext context,
         HomeworkItemInfoViewModel model,
-        Widget child,
+        Widget? child,
       ) {
         return Scaffold(
           appBar: AppBar(
