@@ -1,7 +1,7 @@
-String validateHomeworkItemSubject(String subject) {
-  final String subjectTrimmed = subject.trim();
+String? validateHomeworkItemSubject(String? subject) {
+  final String? subjectTrimmed = subject?.trim();
 
-  if (subjectTrimmed.isEmpty) {
+  if (subjectTrimmed?.isEmpty == true) {
     return "Subject can't be empty.";
   } else if (subjectTrimmed == null) {
     return "Choose a subject.";
