@@ -1,3 +1,5 @@
+import 'package:arrowad_grade_eleven/src/app/models/k_homework.dart';
+import 'package:arrowad_grade_eleven/src/app/models/k_homework_item.dart';
 import 'package:arrowad_grade_eleven/src/ui/views/add_homework/add_homework_view.dart';
 import 'package:arrowad_grade_eleven/src/ui/views/add_homework_item/add_homework_item_view.dart';
 import 'package:arrowad_grade_eleven/src/ui/views/edit_profile/edit_profile_view.dart';
@@ -13,26 +15,29 @@ import 'package:arrowad_grade_eleven/src/ui/views/startup/startup_view.dart';
 import 'package:arrowad_grade_eleven/src/ui/views/teacher_info/teacher_info_view.dart';
 import 'package:arrowad_grade_eleven/src/ui/views/teacher_list/teacher_list_view.dart';
 import 'package:auto_route/auto_route.dart';
+import 'package:flutter/material.dart';
 
-export './router.gr.dart';
+part './router.gr.dart';
 
-@AdaptiveAutoRouter(
+@AutoRouterConfig(
   replaceInRouteName: "View,Route",
-  routes: <AutoRoute<dynamic>>[
-    AutoRoute(page: StartupView, path: "/", initial: true),
-    AutoRoute(page: RegisterView, path: "/register"),
-    AutoRoute(page: HomeView, path: "/home"),
-    AutoRoute(page: MaterialCoveredView, path: "/material-covered"),
-    AutoRoute(page: TeacherInfoView, path: "/teacher/:id"),
-    AutoRoute(page: ScheduleView, path: "/schedule"),
-    AutoRoute(page: TeacherListView, path: "/teacher-list"),
-    AutoRoute(page: LoginView, path: "/login"),
-    AutoRoute(page: HomeworkView, path: "/homework"),
-    AutoRoute(page: AddHomeworkView, path: "/add-homework"),
-    AutoRoute(page: HomeworkItemsView, path: "/homework-items"),
-    AutoRoute(page: HomeworkItemInfoView, path: "/homework-item-info"),
-    AutoRoute(page: AddHomeworkItemView, path: "/add-homework-item"),
-    AutoRoute(page: EditProfileView, path: "/edit-profile"),
-  ],
 )
-class $AppRouter {}
+class AppRouter extends _$AppRouter {
+  @override
+  final List<AutoRoute> routes = [
+    AutoRoute(page: StartupRoute.page, path: "/", initial: true),
+    AutoRoute(page: RegisterRoute.page, path: "/register"),
+    AutoRoute(page: HomeRoute.page, path: "/home"),
+    AutoRoute(page: MaterialCoveredRoute.page, path: "/material-covered"),
+    AutoRoute(page: TeacherInfoRoute.page, path: "/teacher/:id"),
+    AutoRoute(page: ScheduleRoute.page, path: "/schedule"),
+    AutoRoute(page: TeacherListRoute.page, path: "/teacher-list"),
+    AutoRoute(page: LoginRoute.page, path: "/login"),
+    AutoRoute(page: HomeworkRoute.page, path: "/homework"),
+    AutoRoute(page: AddHomeworkRoute.page, path: "/add-homework"),
+    AutoRoute(page: HomeworkItemsRoute.page, path: "/homework-items"),
+    AutoRoute(page: HomeworkItemInfoRoute.page, path: "/homework-item-info"),
+    AutoRoute(page: AddHomeworkItemRoute.page, path: "/add-homework-item"),
+    AutoRoute(page: EditProfileRoute.page, path: "/edit-profile"),
+  ];
+}
