@@ -1,6 +1,5 @@
+import 'package:arrowad_grade_eleven/src/app/router/router.dart';
 import 'package:injectable/injectable.dart';
-
-import 'package:arrowad_grade_eleven/src/app/router/router.gr.dart';
 
 @lazySingleton
 class RouterService {
